+++ common/suites.js
@@ -6,9 +6,6 @@
  */
 
 import {DataIntegrityProof} from '@digitalbazaar/data-integrity';
-import {
-  cryptosuite as ecdsa2019Cryptosuite
-} from '@digitalbazaar/ecdsa-2019-cryptosuite';
 import {cryptosuite as ecdsaRdfc2019Cryptosuite} from
   '@digitalbazaar/ecdsa-rdfc-2019-cryptosuite';
 import {Ed25519Signature2018} from '@digitalbazaar/ed25519-signature-2018';
@@ -20,7 +17,6 @@ import {
 export const SUITES = [
   new Ed25519Signature2018(),
   new Ed25519Signature2020(),
-  new DataIntegrityProof({cryptosuite: ecdsa2019Cryptosuite}),
   new DataIntegrityProof({cryptosuite: ecdsaRdfc2019Cryptosuite}),
   new DataIntegrityProof({cryptosuite: eddsa2022Cryptosuite}),
 ];