+++ CHANGELOG.md
@@ -4,14 +4,18 @@
 
 ### Added
 - New configurable timeout for exchanges `options.recordExpiresDurationMs`.
+- Added support for `ecdsa-rdfc-2019`.
 
 ### Changed
+- Deprecated `exchangeActiveExpirySeconds` config parameter.
 - Improved handling for `EnvelopedVerifiableCredentials` and clarified function
   names.
 - Updated dependencies:
   - `@bedrock/config-yaml@4.3.3` (do not log config on read failure)
   - Update `did-jwt-vc` dependency to latest major version.
   - Updated dependency `@digitalbazaar/vc` to support 2.0 VCDM credentials
+  - `@digitalbazaar/data-integrity`
+  - `@digitalbazaar/did-method-jwk` for authentication proofPurpose
 
 ## 8.6.1 - 2025-01-28